Splitting entry module into several sub-modules
