--- pronto/views/api/entry.py
@@ -1,1932 +1,0 @@
-import re
-
-from cx_Oracle import DatabaseError, IntegrityError, STRING
-from flask import jsonify, request
-
-from pronto import app, executor, db, get_user, xref
-from .annotation import get_citations, insert_citations
-
-
-def _delete_entry(user, dsn, accession):
-    con = db.connect_oracle(user, dsn)
-    cur = con.cursor()
-    try:
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.ENTRY
-            WHERE ENTRY_AC = :1
-            """, (accession,)
-        )
-    except DatabaseError as e:
-        raise e
-    else:
-        con.commit()
-    finally:
-        cur.close()
-        con.close()
-
-
-def update_references(accession):
-    con = db.get_oracle()
-    cur = con.cursor()
-
-    # Get references from the entry-pub table (may need to be updated)
-    cur.execute(
-        """
-        SELECT PUB_ID, ORDER_IN
-        FROM INTERPRO.ENTRY2PUB
-        WHERE ENTRY_AC = :1
-        """, (accession,)
-    )
-    # Keep order_in to now the highest order
-    cur_references = dict(cur.fetchall())
-    # for pub_id, order_in in cur:
-    #
-    #     cur_references[pub_id] = order_in
-    #     # if pub_id in new_references:
-    #     #
-    #     #
-    #     # else:
-    #     #     # Not in annotations any more: can be deleted
-    #     #     to_delete.append(pub_id)
-
-    # Get references from annotations
-    cur.execute(
-        """
-        SELECT TEXT
-        FROM INTERPRO.COMMON_ANNOTATION
-        WHERE ANN_ID IN (
-          SELECT ANN_ID
-          FROM INTERPRO.ENTRY2COMMON
-          WHERE ENTRY_AC = :1
-        )
-        """,
-        (accession, )
-    )
-
-    new_references = set()
-    prog_ref = re.compile(r"\[cite:(PUB\d+)\]", re.I)
-    for row in cur:
-        for m in prog_ref.finditer(row[0]):
-            print(m)
-            new_references.add(m.group(1))
-
-    old_references = set()
-    for pub_id, order_id in cur_references.items():
-        try:
-            new_references.remove(pub_id)
-        except KeyError:
-            # Reference in table is not in annotations: has to be deleted
-            old_references.add(pub_id)
-
-    if old_references:
-        # references to be deleted from ENTRY2PUB
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.ENTRY2PUB
-            WHERE ENTRY_AC = :1
-            AND PUB_ID IN ({})
-            """.format(
-                # i+2: 1-index and +1 for ENTRY_ACC param
-                ','.join([':' + str(i+2) for i in range(len(old_references))])
-            ),
-            (accession,) + tuple(old_references)
-        )
-
-    if new_references:
-        # references to be inserted into ENTRY2PUB
-        start = max(cur_references.values()) + 1 if cur_references else 1
-        params = []
-        for i, pub_id in enumerate(new_references):
-            params.append((accession, start + i, pub_id))
-
-        cur.executemany(
-            """
-            INSERT INTO INTERPRO.ENTRY2PUB (ENTRY_AC, ORDER_IN, PUB_ID) 
-            VALUES (:1, :2, :3)
-            """,
-            params
-        )
-
-    con.commit()
-    cur.close()
-
-
-@app.route("/api/entry/", methods=["PUT"])
-def create_entry():
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    try:
-        entry_type = request.json["type"]
-        entry_name = request.json["name"]
-        entry_descr = request.json["description"]
-    except KeyError:
-        return jsonify({
-            'error': {
-                'title': 'Bad request',
-                'message': 'Invalid or missing parameters.'
-            }
-        }), 400
-
-    if entry_type == 'U':
-        # Unknown type is not allowed
-        return jsonify({
-            "status": False,
-            "title": "Invalid type",
-            "message": "InterPro entries cannot be of type Unknown."
-        }), 400
-
-    con = db.get_oracle()
-    cur = con.cursor()
-
-    # Create the entry
-    accession = cur.var(STRING)
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.ENTRY (ENTRY_AC, ENTRY_TYPE, NAME, SHORT_NAME) 
-            VALUES (INTERPRO.NEW_ENTRY_AC(), :1, :2, :3)
-            RETURNING ENTRY_AC INTO :4
-            """,
-            (entry_type, entry_descr, entry_name, accession)
-        )
-    except DatabaseError as e:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": str(e)
-        }), 401
-
-    # Get the new entry's accession
-    accession = accession.getvalue()
-
-    # Integrate signatures (if any to integrated)
-    signatures = set(request.json.get("signatures", []))
-    if signatures:
-        # Check first they all exist
-        cur.execute(
-            """
-            SELECT METHOD_AC
-            FROM {}.METHOD
-            WHERE METHOD_AC IN ({})
-            """.format(
-                app.config["DB_SCHEMA"],
-                ','.join([':'+str(i+1) for i in range(len(signatures))])
-            ),
-            tuple(signatures)
-        )
-        existing_signatures = {row[0] for row in cur}
-
-        for acc in signatures:
-            if acc not in existing_signatures:
-                cur.close()
-                return jsonify({
-                    "status": False,
-                    "title": "Invalid signature",
-                    "message": "<strong>{}</strong> is not a valid "
-                               "member database accession.".format(acc)
-                }), 401
-
-        try:
-            cur.execute(
-                """
-                DELETE FROM INTERPRO.ENTRY2METHOD
-                WHERE METHOD_AC IN ({})
-                """.format(
-                    ','.join(
-                        [':' + str(i + 1) for i in range(len(signatures))])
-                ),
-                tuple(signatures)
-            )
-
-            cur.executemany(
-                """
-                INSERT INTO INTERPRO.ENTRY2METHOD (
-                  ENTRY_AC, METHOD_AC, EVIDENCE
-                )
-                VALUES (:1, :2, 'MAN')
-                """,
-                [(accession, acc) for acc in signatures]
-            )
-        except DatabaseError as e:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Database error",
-                "message": str(e)
-            }), 401
-
-    con.commit()
-    cur.close()
-    return jsonify({
-        "status": True,
-        "accession": accession
-    }), 200
-
-
-@app.route("/api/entry/<accession>/")
-def get_entry(accession):
-    if executor.has(accession):
-        return jsonify(None), 404
-
-    cur = db.get_oracle().cursor()
-    cur.execute(
-        """
-        SELECT
-          E.NAME,
-          E.SHORT_NAME,
-          E.ENTRY_TYPE,
-          ET.ABBREV,
-          E.CHECKED,
-          E.TIMESTAMP
-        FROM INTERPRO.ENTRY E
-        INNER JOIN INTERPRO.CV_ENTRY_TYPE ET 
-          ON E.ENTRY_TYPE = ET.CODE
-        WHERE E.ENTRY_AC = :1
-        """,
-        (accession,)
-    )
-
-    row = cur.fetchone()
-    cur.close()
-    if row:
-        entry = {
-            "accession": accession,
-            "name": row[0],
-            "short_name": row[1],
-            "type": {
-                "code": row[2],
-                "name": row[3].replace('_', ' ')
-            },
-            "is_checked": row[4] == 'Y',
-            "last_modification": row[5].strftime("%d %b %Y")
-        }
-        return jsonify(entry), 200
-    else:
-        return jsonify(None), 404
-
-
-@app.route("/api/entry/<accession>/", methods=["POST"])
-def update_entry(accession):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    try:
-        name = request.form["name"].strip()
-        assert 0 < len(name) <= 30
-    except (AssertionError, KeyError):
-        return jsonify({
-            "status": False,
-            "title": "Invalid or missing parameter",
-            "message": "'name' must be between 1 and 30 characters long."
-        }), 400
-
-    try:
-        description = request.form["description"].strip()
-        assert 0 < len(description) <= 100
-    except (AssertionError, KeyError):
-        return jsonify({
-            "status": False,
-            "title": "Invalid or missing parameter",
-            "message": "'description' must be between 1 "
-                       "and 100 characters long."
-        }), 400
-
-    try:
-        _type = request.form["type"].strip()
-    except KeyError:
-        return jsonify({
-            "status": False,
-            "title": "Missing parameter",
-            "message": "'type' must be provided."
-        }), 400
-
-    try:
-        is_checked = int(request.form["checked"].strip())
-    except (KeyError, ValueError):
-        return jsonify({
-            "status": False,
-            "title": "Invalid or missing parameter",
-            "message": "'checked' must be an integer."
-        }), 400
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT ENTRY_TYPE
-        FROM INTERPRO.ENTRY
-        WHERE ENTRY_AC = :1
-        """, (accession,)
-    )
-    row = cur.fetchone()
-    if not row:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "{} is not a "
-                       "valid InterPro accession.".format(accession)
-        }), 404
-    current_type = row[0]
-    if _type == 'U':
-        # Unknown type is not allowed
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid type",
-            "message": "InterPro entries cannot be of type Unknown."
-        }), 400
-    elif _type != current_type:
-        # Check that we do not have relationships with other entries
-        cur.execute(
-            """
-            SELECT COUNT(*)
-            FROM INTERPRO.ENTRY2ENTRY
-            WHERE ENTRY_AC = :acc
-            OR PARENT_AC = :acc
-            """,
-            dict(acc=accession)
-        )
-        if cur.fetchone()[0]:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Cannot update type",
-                "message": "{} cannot have its type changed because "
-                           "it has InterPro relationships.".format(accession)
-            }), 400
-
-    if is_checked:
-        cur.execute(
-            """
-            SELECT COUNT(*)
-            FROM INTERPRO.ENTRY2METHOD
-            WHERE ENTRY_AC = :1
-            """, (accession,)
-        )
-        if not cur.fetchone()[0]:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Cannot check entry",
-                "message": "{} cannot be checked because it does not have "
-                           "any signatures.".format(accession)
-            }), 400
-
-        cur.execute(
-            """
-            SELECT COUNT(*)
-            FROM INTERPRO.ENTRY2COMMON
-            WHERE ENTRY_AC = :1
-            """, (accession,)
-        )
-        if not cur.fetchone()[0]:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Cannot check entry",
-                "message": "{} cannot be checked because it does not have "
-                           "any annotations.".format(accession)
-            }), 400
-    try:
-        cur.execute(
-            """
-            UPDATE INTERPRO.ENTRY
-            SET 
-              ENTRY_TYPE = :1, 
-              NAME = :2,
-              CHECKED = :3,
-              SHORT_NAME = :4
-            WHERE ENTRY_AC = :5
-            """,
-            (_type, description, 'Y' if is_checked else 'N', name, accession)
-        )
-    except DatabaseError as e:
-        print(e)
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not update {}.".format(accession)
-        }), 500
-    else:
-        con.commit()
-        return jsonify({
-            "status": True
-        }), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<accession>/", methods=["DELETE"])
-def delete_entry(accession):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT COUNT(*)
-        FROM INTERPRO.ENTRY
-        WHERE ENTRY_AC = :1
-        """, (accession,)
-    )
-    if not cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "{} is not a "
-                       "valid InterPro accession.".format(accession)
-        }), 404
-
-    cur.execute(
-        """
-        SELECT COUNT(*)
-        FROM INTERPRO.ENTRY2METHOD
-        WHERE ENTRY_AC = :1
-        """, (accession,)
-    )
-    n_signatures = cur.fetchone()[0]
-    cur.close()
-
-    if n_signatures:
-        return jsonify({
-            "status": False,
-            "title": "Cannot delete entry",
-            "message": "{} cannot be deleted because "
-                       "it has one or more signatures.".format(accession)
-        }), 400
-    else:
-        dsn = app.config["ORACLE_DB"]["dsn"]
-        executor.submit(accession, _delete_entry, user, dsn, accession)
-        return jsonify({"status": True}), 202
-
-
-@app.route('/api/entry/<acc>/annotation/<ann_id>/', methods=["DELETE"])
-def unlink_annotation(acc, ann_id):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-
-    try:
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.ENTRY2COMMON
-            WHERE ENTRY_AC = :1 AND ANN_ID = :2 
-            """,
-            (acc, ann_id)
-        )
-    except DatabaseError:
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not unlink annotation."
-        }), 500
-    else:
-        con.commit()
-        return jsonify({
-            "status": True
-        }), 200
-    finally:
-        cur.close()
-        update_references(acc)
-
-
-@app.route('/api/entry/<acc>/annotation/<ann_id>/', methods=["PUT"])
-def link_annotation(acc, ann_id):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT MAX(ORDER_IN)
-        FROM INTERPRO.ENTRY2COMMON
-        WHERE ENTRY_AC = :1
-        """, (acc,)
-    )
-    order_in = cur.fetchone()[0]
-    if order_in is None:
-        order_in = 0
-    else:
-        order_in += 1
-
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.ENTRY2COMMON (ENTRY_AC, ANN_ID, ORDER_IN)
-            VALUES (:1, :2, :3) 
-            """,
-            (acc, ann_id, order_in)
-        )
-    except (DatabaseError, IntegrityError):
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not link annotation."
-        }), 500
-    else:
-        con.commit()
-        return jsonify({
-            "status": True
-        }), 200
-    finally:
-        update_references(acc)
-        cur.close()
-
-
-@app.route('/api/entry/<acc>/annotation/<ann_id>/order/<x>/', methods=["POST"])
-def reorder_annotation(acc, ann_id, x):
-    try:
-        x = int(x)
-    except ValueError:
-        return jsonify({
-            "status": False,
-            "title": "Invalid parameters",
-            "message": "Expected an integer."
-        }), 400
-
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-
-    # Get the position of every annotation
-    cur.execute(
-        """
-        SELECT ANN_ID, ORDER_IN
-        FROM INTERPRO.ENTRY2COMMON
-        WHERE ENTRY_AC = :1
-        """,
-        (acc,)
-    )
-    annotations = dict(cur.fetchall())
-
-    if ann_id not in annotations:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid parameters",
-            "message": "{} is not linked by {}".format(ann_id, acc)
-        }), 400
-
-    max_order = max(annotations.values())
-    min_order = min(annotations.values())
-
-    # Annotations in the current order
-    annotations = sorted(annotations, key=lambda k: annotations[k])
-
-    # Current index of the annotation to move
-    idx = annotations.index(ann_id)
-
-    # New index, forced to stay in the range [0, num_annotations-1]
-    new_idx = min(max(0, idx + x), len(annotations)-1)
-
-    if idx == new_idx:
-        # Nothing to change
-        cur.close()
-        return jsonify({
-            "status": True
-        }), 200
-
-    # Remove the annotation, then insert it at its new index
-    annotations.pop(idx)
-    annotations.insert(new_idx, ann_id)
-
-    """
-    We would like to use the range [0, num_annotations[
-    but that is only possible if max_order > min_order >= num_annotations.
-    If it is not, we are using the range [max_order+1, max_order+1+num_annotations[
-    """
-    if max_order > min_order >= len(annotations):
-        offset = 0
-    else:
-        offset = max_order + 1
-
-    for i, ann_id in enumerate(annotations):
-        cur.execute(
-            """
-            UPDATE INTERPRO.ENTRY2COMMON
-            SET ORDER_IN = :1
-            WHERE ENTRY_AC = :2 AND ANN_ID = :3
-            """, (i + offset, acc, ann_id)
-        )
-
-    con.commit()
-    cur.close()
-
-    return jsonify({
-        "status": True
-    }), 200
-
-
-@app.route('/api/entry/<accession>/references/')
-def get_suppl_references(accession):
-    cur = db.get_oracle().cursor()
-    cur.execute(
-        """
-        SELECT 
-          S.PUB_ID, C.TITLE, C.YEAR, C.VOLUME, C.RAWPAGES, C.DOI_URL,
-          C.PUBMED_ID, C.ISO_JOURNAL, C.MEDLINE_JOURNAL, C.AUTHORS
-        FROM INTERPRO.SUPPLEMENTARY_REF S
-        INNER JOIN INTERPRO.CITATION C
-        ON S.PUB_ID = C.PUB_ID
-        WHERE S.ENTRY_AC = :1      
-        """, (accession,)
-    )
-    references = []
-    for row in cur:
-        references.append({
-            "id": row[0],
-            "title": row[1],
-            "year": row[2],
-            "volume": row[3],
-            "pages": row[4],
-            "doi": row[5],
-            "pmid": row[6],
-            "journal": row[7] if row[7] else row[8],
-            "authors": row[9]
-        })
-    cur.close()
-    return jsonify(references)
-
-
-@app.route('/api/entry/<accession>/annotations/')
-def get_annotations_new(accession):
-    cur = db.get_oracle().cursor()
-
-    # Annotations
-    cur.execute(
-        """
-        SELECT A.ANN_ID, A.TEXT, A.COMMENTS, S.CT
-        FROM INTERPRO.COMMON_ANNOTATION A
-        INNER JOIN INTERPRO.ENTRY2COMMON E ON A.ANN_ID = E.ANN_ID
-        LEFT OUTER JOIN (
-          SELECT ANN_ID, COUNT(*) CT
-          FROM INTERPRO.ENTRY2COMMON
-          GROUP BY ANN_ID
-        ) S ON A.ANN_ID = S.ANN_ID
-        WHERE E.ENTRY_AC = :1
-        ORDER BY E.ORDER_IN
-        """,
-        (accession,)
-    )
-
-    annotations = []
-    prog_ref = re.compile(r"\[([a-z]+):([a-z0-9\-.]+)]", re.I)
-
-    for ann_id, text, comment, n_entries in cur:
-        ext_refs = {}
-
-        for match in prog_ref.finditer(text):
-            ref_db, ref_id = match.groups()
-            ref_db = ref_db.upper()
-            base_url = xref.find_xref(ref_db)
-            if base_url:
-                whole_match = match.group()
-
-                ext_refs[whole_match] = {
-                    "match": whole_match,
-                    "id": ref_id,
-                    "url": base_url.format(ref_id)
-                }
-
-        annotations.append({
-            "id": ann_id,
-            "text": text,
-            "comment": comment,
-            "num_entries": n_entries,
-            "xrefs": list(ext_refs.values())
-        })
-
-    cur.execute(
-        """
-        SELECT 
-          E.PUB_ID, C.TITLE, C.YEAR, C.VOLUME, C.RAWPAGES, C.DOI_URL,
-          C.PUBMED_ID, C.ISO_JOURNAL, C.MEDLINE_JOURNAL, C.AUTHORS
-        FROM INTERPRO.ENTRY2PUB E
-        INNER JOIN INTERPRO.CITATION C
-        ON E.PUB_ID = C.PUB_ID
-        WHERE E.ENTRY_AC = :1
-        """, (accession,)
-    )
-
-    references = {}
-    for row in cur:
-        pub_id = row[0]
-        references[pub_id] = {
-            "id": pub_id,
-            "title": row[1],
-            "year": row[2],
-            "volume": row[3],
-            "pages": row[4],
-            "doi": row[5],
-            "pmid": row[6],
-            "journal": row[7] if row[7] else row[8],
-            "authors": row[9]
-        }
-
-    cur.close()
-    return jsonify({
-        "annotations": annotations,
-        "references": references
-    })
-
-
-#@app.route('/api/entry/<accession>/annotations/')
-def get_annotations(accession):
-    cur = db.get_oracle().cursor()
-
-    # annotations
-    cur.execute(
-        """
-        SELECT A.ANN_ID, A.TEXT, A.COMMENTS, S.CT
-        FROM INTERPRO.COMMON_ANNOTATION A
-        INNER JOIN INTERPRO.ENTRY2COMMON E ON A.ANN_ID = E.ANN_ID
-        LEFT OUTER JOIN (
-          SELECT ANN_ID, COUNT(*) CT
-          FROM INTERPRO.ENTRY2COMMON
-          GROUP BY ANN_ID
-        ) S ON A.ANN_ID = S.ANN_ID
-        WHERE E.ENTRY_AC = :1
-        ORDER BY E.ORDER_IN
-        """,
-        (accession,)
-    )
-
-    annotations = []
-    text_references = set()
-    xrefs = []
-    prog_ref = r'<cite\s+id="(PUB\d+)"\s*/>'
-    prog_xref = r'<dbxref\s+db\s*=\s*"(\w+)"\s+id\s*=\s*"([\w\.\-]+)"\s*\/>'
-    prog_taxa = r'<taxon\s+tax_id="(\d+)">([^<]+)</taxon>'
-
-    for ann_id, text, comment, count in cur:
-        for m in re.finditer(prog_ref, text):
-            text_references.add(m.group(1))
-
-        # Find cross-references
-        for m in re.finditer(prog_xref, text):
-            match = m.group(0)
-            dbcode = m.group(1).upper()
-            xref_id = m.group(2)
-
-            url = xref.find_xref(dbcode)
-            if url:
-                url = url.format(xref_id)
-                xrefs.append({
-                    "tag": match,
-                    "url": url,
-                    "id": xref_id
-                })
-            else:
-                # TODO: what do we do with unknown cross-references?
-                continue
-
-        for m in re.finditer(prog_taxa, text):
-            xrefs.append({
-                "tag": m.group(0),
-                "url": "https://www.uniprot.org/taxonomy/{}/".format(
-                    m.group(1)
-                ),
-                "id": m.group(2)
-            })
-
-        annotations.append({
-            "id": ann_id,
-            "text": text,
-            "comment": comment,
-            "num_entries": count
-        })
-
-    """
-    Get references from ref tables
-    We consider that entries in ENTRY2PUB are update automatically
-        (i.e. when an annotation is updated, linked, unlinked)
-    and entries in SUPPLEMENTARY_REF are created by curators (i.e. deletable)
-    """
-    cur.execute(
-        """
-        SELECT 
-          PUB.PUB_ID, PUB.DELETABLE, 
-          C.TITLE, C.YEAR, C.VOLUME, C.RAWPAGES, C.DOI_URL,
-          C.PUBMED_ID, C.ISO_JOURNAL, C.MEDLINE_JOURNAL, C.AUTHORS
-        FROM (
-          SELECT PUB_ID, 'N' AS DELETABLE
-          FROM INTERPRO.ENTRY2PUB
-          WHERE ENTRY_AC = :acc
-          UNION ALL
-          SELECT SUPPLEMENTARY_REF.PUB_ID, 'Y' AS DELETABLE
-          FROM INTERPRO.SUPPLEMENTARY_REF
-          WHERE ENTRY_AC = :acc        
-        ) PUB
-        INNER JOIN INTERPRO.CITATION C
-          ON PUB.PUB_ID = C.PUB_ID
-        """,
-        dict(acc=accession)
-    )
-
-    references = {}
-    columns = ("id", "deletable",
-               "title", "year", "volume", "pages", "doi",
-               "pmid", "journal_iso", "journal_medline", "authors")
-    for row in cur:
-        pub_id = row[0]
-        pub = dict(zip(columns, row))
-        pub["deletable"] = pub["deletable"] == 'Y'
-
-        if pub_id not in references:
-            references[pub_id] = pub
-        elif pub["deletable"] and not references[pub_id]["deletable"]:
-            references[pub_id] = pub
-
-    missing_references = text_references - set(references)
-    if missing_references:
-        # Some text references are (somehow) not in the entry-pub table
-        in_stmt = [':' + str(i+1) for i in range(len(missing_references))]
-        cur.execute(
-            """
-            SELECT
-              PUB_ID, 'N', 
-              TITLE, YEAR, VOLUME, RAWPAGES, DOI_URL,
-              PUBMED_ID, ISO_JOURNAL, MEDLINE_JOURNAL, AUTHORS
-            FROM INTERPRO.CITATION
-            WHERE PUB_ID IN ({})
-            """.format(','.join(in_stmt)),
-            tuple(missing_references)
-        )
-
-        for row in cur:
-            references[row[0]] = dict(zip(columns, row))
-
-    cur.close()
-
-    # Select journal (default: ISO, fallback to MEDLINE)
-    for ref, pub in references.items():
-        journal_iso = pub.pop("journal_iso")
-        journal_med = pub.pop("journal_medline")
-        if journal_iso:
-            references[ref]["journal"] = journal_iso
-        else:
-            references[ref]["journal"] = journal_med
-
-    return jsonify({
-        "annotations": annotations,
-        "references": references,
-        "cross_references": xrefs
-    }), 200
-
-
-@app.route("/api/entry/<accession>/check/", methods=["POST"])
-def check_entry(accession):
-    content = request.get_json()
-    try:
-        is_checked = bool(int(content["checked"]))
-    except (KeyError, ValueError):
-        return jsonify({
-            "status": False,
-            "title": "Bad request",
-            "message": "Invalid or missing parameters."
-        }), 400
-
-    user = get_user()
-
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-
-    if is_checked:
-        cur.execute(
-            """
-            SELECT COUNT(*)
-            FROM INTERPRO.ENTRY2METHOD
-            WHERE ENTRY_AC = :1
-            """, (accession,)
-        )
-        if not cur.fetchone()[0]:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Cannot check entry",
-                "message": "{} cannot be checked because it does not have "
-                           "any signatures.".format(accession)
-            }), 400
-
-        cur.execute(
-            """
-            SELECT COUNT(*)
-            FROM INTERPRO.ENTRY2COMMON
-            WHERE ENTRY_AC = :1
-            """, (accession,)
-        )
-        if not cur.fetchone()[0]:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Cannot check entry",
-                "message": "{} cannot be checked because it does not have "
-                           "any annotations.".format(accession)
-            }), 400
-
-    try:
-        cur.execute(
-            """
-            UPDATE INTERPRO.ENTRY
-            SET CHECKED = :1
-            WHERE ENTRY_AC = :2
-            """,
-            ('Y' if is_checked else 'N', accession)
-        )
-    except DatabaseError as e:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not update {}: {}".format(accession, e)
-        }), 500
-    else:
-        con.commit()
-        cur.close()
-        return jsonify({
-            "status": True,
-            "message": None
-        }), 200
-
-
-@app.route("/api/entry/<parent_acc>/child/<child_acc>/", methods=["PUT"])
-def add_child_relationship(parent_acc, child_acc):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT ENTRY_AC, ENTRY_TYPE
-        FROM INTERPRO.ENTRY 
-        WHERE ENTRY_AC = :1 OR ENTRY_AC = :2
-        """, (parent_acc, child_acc)
-    )
-    entries = dict(cur.fetchall())
-
-    if parent_acc not in entries:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "{} is not "
-                       "a valid InterPro accession.".format(parent_acc)
-        }), 400
-    elif child_acc not in entries:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "{} is not "
-                       "a valid InterPro accession.".format(child_acc)
-        }), 400
-    elif entries[parent_acc] != entries[child_acc]:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid relationship",
-            "message": "{} and {} do not have the same type.".format(
-                parent_acc, child_acc
-            )
-        }), 400
-
-    cur.execute(
-        """
-        SELECT COUNT(*) 
-        FROM INTERPRO.ENTRY2ENTRY
-        WHERE PARENT_AC = :1 AND ENTRY_AC = :2
-        """, (parent_acc, child_acc)
-    )
-    if cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": True
-        }), 200
-
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.ENTRY2ENTRY (ENTRY_AC, PARENT_AC, RELATION)
-            VALUES (:1, :2, :3)
-            """, (child_acc, parent_acc, "TY")
-        )
-    except (DatabaseError, IntegrityError):
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not link {} to {}.".format(parent_acc, child_acc)
-        }), 500
-    else:
-        con.commit()
-        return jsonify({
-            "status": True
-        }), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<accession>/comment/", methods=["PUT"])
-def add_entry_comment(accession):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    content = request.get_json()
-    text = content.get("text", "")
-    if len(text) < 3:
-        return jsonify({
-            "status": False,
-            "title": "Comment to short",
-            "message": "Comment must be at least three characters long."
-        }), 400
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT MAX(ID)
-        FROM INTERPRO.ENTRY_COMMENT
-        """
-    )
-    max_id = cur.fetchone()[0]
-    next_id = max_id + 1 if max_id else 1
-
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.ENTRY_COMMENT (
-              ID, ENTRY_AC, USERNAME, VALUE
-            )
-            VALUES (:1, :2, :3, :4)
-            """,
-            (next_id, accession, user["username"], text)
-        )
-    except IntegrityError as e:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not add comment "
-                       "for {}: {}".format(accession, e)
-        }), 500
-    else:
-        con.commit()
-        cur.close()
-        return jsonify({
-            "status": True,
-            "message": None
-        }), 200
-
-
-@app.route("/api/entry/<accession>/comment/<_id>/", methods=["DELETE"])
-def delete_entry_comment(accession, _id):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    try:
-        cur.execute(
-            """
-            UPDATE INTERPRO.ENTRY_COMMENT 
-            SET STATUS = 'N' 
-            WHERE ENTRY_AC = :1 AND ID = :2
-            """,
-            (accession, _id)
-        )
-    except DatabaseError as e:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not delete comment "
-                       "for {}: {}".format(accession, e)
-        }), 500
-    else:
-        con.commit()
-        cur.close()
-        return jsonify({
-            "status": True,
-            "message": None
-        }), 200
-
-
-@app.route("/api/entry/<accession>/comments/")
-def get_entry_comments(accession):
-    try:
-        n = int(request.args["max"])
-    except (KeyError, ValueError):
-        n = 0
-
-    cur = db.get_oracle().cursor()
-    cur.execute(
-        """
-        SELECT C.ID, C.VALUE, C.CREATED_ON, C.STATUS, U.NAME
-        FROM INTERPRO.ENTRY_COMMENT C
-        INNER JOIN INTERPRO.USER_PRONTO U ON C.USERNAME = U.USERNAME
-        WHERE C.ENTRY_AC = :1
-        ORDER BY C.CREATED_ON DESC
-        """,
-        (accession,)
-    )
-
-    comments = [
-        {
-            "id": row[0],
-            "text": row[1],
-            "date": row[2].strftime("%Y-%m-%d %H:%M:%S"),
-            "status": row[3] == "Y",
-            "author": row[4],
-        } for row in cur
-    ]
-    cur.close()
-    n_comments = len(comments)
-
-    if n:
-        comments = comments[:n]
-
-    for c in comments:
-        c["text"] = re.sub(
-            r"#(\d+)",
-            r'<a href="https://github.com/geneontology/go-annotation/issues/\1">#\1</a>',
-            c["text"]
-        )
-
-    return jsonify({
-        "count": n_comments,
-        "comments": comments
-    })
-
-
-@app.route("/api/entry/<accession>/go/")
-def get_entry_go_terms(accession):
-    cur = db.get_oracle().cursor()
-
-    cur.execute(
-        """
-        SELECT 
-          T.GO_ID, T.NAME, T.CATEGORY, T.DEFINITION, 
-          T.IS_OBSOLETE, T.REPLACED_BY
-        FROM INTERPRO.INTERPRO2GO I
-          INNER JOIN {}.TERM T ON I.GO_ID = T.GO_ID
-        WHERE I.ENTRY_AC = :1
-        ORDER BY T.GO_ID
-        """.format(app.config["DB_SCHEMA"]),
-        (accession,)
-    )
-
-    terms = []
-    for row in cur:
-        terms.append({
-            "id": row[0],
-            "name": row[1],
-            "category": row[2],
-            "definition": row[3],
-            "is_obsolete": row[4] == 'Y',
-            "secondary": row[5] is not None
-        })
-
-    cur.close()
-
-    return jsonify(terms), 200
-
-
-@app.route("/api/entry/<accession>/go/<term_id>/", methods=['DELETE'])
-def delete_go_term(accession, term_id):
-    user = get_user()
-
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    try:
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.INTERPRO2GO
-            WHERE ENTRY_AC = :1 AND GO_ID = :2
-            """,
-            (accession, term_id)
-        )
-    except IntegrityError:
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not delete GO term {} "
-                       "from InterPro entry {}".format(accession, term_id)
-        }), 500
-    else:
-        # row_count = cur.rowcount  # TODO: check that row_count == 1?
-        con.commit()
-        return jsonify({
-            "status": True,
-            "title": None,
-            "message": None
-        }), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<accession>/go/<term_id>/", methods=["PUT"])
-def add_go_term(accession, term_id):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT COUNT(*) 
-        FROM INTERPRO.ENTRY 
-        WHERE ENTRY_AC = :1
-        """, (accession,)
-    )
-    if not cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "{} is not a valid InterPro accession.".format(accession)
-        }), 400
-
-    cur.execute(
-        """
-        SELECT COUNT(*) 
-        FROM {}.TERM 
-        WHERE GO_ID = :1
-        """.format(app.config["DB_SCHEMA"]), (term_id,)
-    )
-    if not cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid GO term",
-            "message": "{} is not a valid GO term ID".format(term_id)
-        }), 401
-
-    cur.execute(
-        """
-        SELECT COUNT(*)
-        FROM INTERPRO.INTERPRO2GO
-        WHERE ENTRY_AC = :1 AND GO_ID = :2
-        """, (accession, term_id)
-    )
-    if cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": True
-        }), 200
-
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.INTERPRO2GO (ENTRY_AC, GO_ID, SOURCE)
-            VALUES (:1, :2, :3)
-            """, (accession, term_id, "MANU")
-        )
-    except (DatabaseError, IntegrityError):
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not add {} to {}".format(term_id, accession)
-        }), 500
-    else:
-        con.commit()
-        return jsonify({
-            "status": True
-        }), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<accession>/reference/<int:pmid>/", methods=["PUT"])
-def link_reference(accession, pmid):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT PUB_ID
-        FROM INTERPRO.CITATION
-        WHERE PUBMED_ID = :1
-        """, (pmid,)
-    )
-    row = cur.fetchone()
-    if row:
-        pub_id = row[0]
-    else:
-        citations = get_citations(cur, [pmid])
-        if pmid not in citations:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Invalid reference",
-                "message": "<strong>{}</strong> is not "
-                           "a valid PubMed ID".format(pmid)
-            }), 400
-
-        error = insert_citations(cur, citations)
-        if error:
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Database error",
-                "message": error
-            }), 500
-
-        pub_id = citations[pmid]
-
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.SUPPLEMENTARY_REF (ENTRY_AC, PUB_ID) 
-            VALUES (:1, :2)
-            """, (accession, pub_id)
-        )
-    except DatabaseError:
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not add {} to {}.".format(pmid, accession)
-        }), 500
-    else:
-        con.commit()
-        return jsonify({"status": True}), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<accession>/reference/<pub_id>/", methods=["DELETE"])
-def unlink_reference(accession, pub_id):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    try:
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.SUPPLEMENTARY_REF
-            WHERE ENTRY_AC = :1 AND PUB_ID = :2
-            """, (accession, pub_id)
-        )
-    except DatabaseError:
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not unlink reference."
-        }), 500
-    else:
-        if cur.rowcount:
-            con.commit()
-        return jsonify({"status": True}), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<acc1>/relationship/<acc2>/", methods=['DELETE'])
-def delete_relationship(acc1, acc2):
-    user = get_user()
-
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    try:
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.ENTRY2ENTRY
-            WHERE (PARENT_AC = :acc1 AND ENTRY_AC = :acc2)
-            OR (PARENT_AC = :acc2 AND ENTRY_AC = :acc1)
-            """,
-            dict(acc1=acc1, acc2=acc2)
-        )
-    except IntegrityError:
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not delete unlink {} and {}".format(acc1, acc2)
-        }), 500
-    else:
-        # row_count = cur.rowcount  # TODO: check that row_count == 1?
-        con.commit()
-        return jsonify({
-            "status": True,
-            "title": None,
-            "message": None
-        }), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<accession>/relationships/")
-def get_entry_relationships(accession):
-    cur = db.get_oracle().cursor()
-
-    cur.execute(
-        """
-        SELECT 
-          R.PARENT_AC, E1.NAME, E1.ENTRY_TYPE, 
-          R.ENTRY_AC, E2.NAME, E2.ENTRY_TYPE
-        FROM (
-            SELECT PARENT_AC, ENTRY_AC
-            FROM INTERPRO.ENTRY2ENTRY
-            START WITH ENTRY_AC = :accession
-            CONNECT BY PRIOR PARENT_AC = ENTRY_AC
-            UNION ALL
-            SELECT PARENT_AC, ENTRY_AC
-            FROM INTERPRO.ENTRY2ENTRY
-            START WITH PARENT_AC = :accession
-            CONNECT BY PRIOR ENTRY_AC = PARENT_AC        
-        ) R
-        INNER JOIN INTERPRO.ENTRY E1 
-          ON R.PARENT_AC = E1.ENTRY_AC
-        INNER JOIN INTERPRO.ENTRY E2 
-          ON R.ENTRY_AC = E2.ENTRY_AC
-        """,
-        dict(accession=accession)
-    )
-
-    parents = {}
-    hierarchy = {}
-    for row in cur:
-        parent_acc = row[0]
-        parent_name = row[1]
-        parent_type = row[2]
-        child_acc = row[3]
-        child_name = row[4]
-        child_type = row[5]
-
-        parents[child_acc] = parent_acc
-
-        if parent_acc in parents:
-            _parents = []
-            root_acc = parent_acc
-            while root_acc in parents:
-                root_acc = parents[parent_acc]
-                _parents.append(root_acc)
-
-            children = {}
-            for acc in reversed(_parents):
-                if not children:
-                    children = hierarchy[acc]["children"]
-                else:
-                    children = children[acc]["children"]
-
-            node = children[parent_acc]
-        elif parent_acc in hierarchy:
-            node = hierarchy[parent_acc]
-        else:
-            node = hierarchy[parent_acc] = {
-                "accession": parent_acc,
-                "name": parent_name,
-                "type": parent_type,
-                "children": {},
-                "deletable": child_acc == accession
-            }
-
-        if child_acc in hierarchy:
-            child = hierarchy.pop(child_acc)
-        else:
-            child = {
-                "accession": child_acc,
-                "name": child_name,
-                "type": child_type,
-                "children": {},
-                "deletable": parent_acc == accession
-            }
-
-        node["children"][child_acc] = child
-
-    cur.close()
-    return jsonify(hierarchy), 200
-
-
-@app.route("/api/entry/<e_acc>/signature/<s_acc>/", methods=['PUT'])
-def integrate_signature(e_acc, s_acc):
-    # If True, move a signature from one entry to another
-    move_signature = request.form.get("confirm") is not None
-
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT COUNT(*)
-        FROM INTERPRO.ENTRY
-        WHERE ENTRY_AC = :1
-        """,
-        (e_acc,)
-    )
-    if not cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "<strong>{}</strong> is not "
-                       "a valid InterPro accession".format(s_acc)
-        }), 401
-
-    cur.execute(
-        """
-        SELECT M.METHOD_AC, EM.ENTRY_AC
-        FROM {}.METHOD M
-        LEFT OUTER JOIN INTERPRO.ENTRY2METHOD EM 
-          ON M.METHOD_AC = EM.METHOD_AC
-        WHERE UPPER(M.METHOD_AC) = :acc 
-        OR UPPER(M.NAME) = :acc
-        """.format(app.config["DB_SCHEMA"]),
-        dict(acc=s_acc.upper())
-    )
-    row = cur.fetchone()
-    if not row:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid signature",
-            "message": "<strong>{}</strong> is not a valid member database "
-                       "accession or name.".format(s_acc)
-        }), 401
-
-    """
-    Set to True if:
-        - the signature is already integrated into another entry
-        - we move the signature from this entry to `e_acc`
-        - the signature was the *last* signature of the other entry
-                -> the other entry must be unchecked 
-    """
-    unchecked = False
-
-    s_acc, in_entry_acc = row
-    if in_entry_acc == e_acc:
-        # Already integrated in this entry: do noting
-        cur.close()
-        return jsonify({"status": True}), 200
-    elif in_entry_acc:
-        # Already integrated in an other entry
-
-        if move_signature:
-            try:
-                cur.execute(
-                    """
-                    DELETE FROM INTERPRO.ENTRY2METHOD
-                    WHERE ENTRY_AC = :1 AND METHOD_AC = :2
-                    """,
-                    (in_entry_acc, s_acc)
-                )
-            except DatabaseError:
-                cur.close()
-                return jsonify({
-                    "status": False,
-                    "title": "Database error",
-                    "message": "Could not unintegrated "
-                               "{} from {}".format(s_acc, in_entry_acc)
-                }), 500
-
-            cur.execute(
-                """
-                SELECT COUNT(*)
-                FROM INTERPRO.ENTRY2METHOD
-                WHERE ENTRY_AC  =:1
-                """,
-                (in_entry_acc,)
-            )
-
-            if not cur.fetchone()[0]:
-                # Need to uncheck the entry
-                try:
-                    cur.execute(
-                        """
-                        UPDATE INTERPRO.ENTRY
-                        SET CHECKED = 'N'
-                        WHERE ENTRY_AC = :1
-                        """,
-                        (in_entry_acc,)
-                    )
-                except DatabaseError:
-                    cur.close()
-                    return jsonify({
-                        "status": False,
-                        "title": "Database error",
-                        "message": "Could not uncheck {}".format(in_entry_acc)
-                    }), 500
-                else:
-                    unchecked = True
-        else:
-            # Ask for confirmation
-            cur.close()
-            return jsonify({
-                "status": True,
-                "signature": s_acc,
-                "entry": in_entry_acc
-            }), 200
-
-    try:
-        cur.execute(
-            """
-            INSERT INTO INTERPRO.ENTRY2METHOD (ENTRY_AC, METHOD_AC, EVIDENCE) 
-            VALUES (:1, :2, 'MAN')
-            """,
-            (e_acc, s_acc)
-        )
-    except DatabaseError:
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not integrated "
-                       "{} into {}".format(s_acc, e_acc)
-        }), 500
-    else:
-        con.commit()
-        if unchecked:
-            res = {
-                "status": True,
-                "unchecked": unchecked,
-                "entry": in_entry_acc
-            }
-        else:
-            res = {
-                "status": True
-            }
-
-        return jsonify(res), 200
-    finally:
-        cur.close()
-
-
-@app.route("/api/entry/<e_acc>/signature/<s_acc>/", methods=['DELETE'])
-def unintegrate_signature(e_acc, s_acc):
-    user = get_user()
-    if not user:
-        return jsonify({
-            "status": False,
-            "title": "Access denied",
-            "message": 'Please <a href="/login/">log in</a> '
-                       'to perform this operation.'
-        }), 401
-
-    con = db.get_oracle()
-    cur = con.cursor()
-    cur.execute(
-        """
-        SELECT COUNT(*)
-        FROM INTERPRO.ENTRY
-        WHERE ENTRY_AC = :1
-        """,
-        (e_acc,)
-    )
-    if not cur.fetchone()[0]:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Invalid entry",
-            "message": "<strong>{}</strong> is not "
-                       "a valid InterPro accession".format(s_acc)
-        }), 401
-
-    try:
-        cur.execute(
-            """
-            DELETE FROM INTERPRO.ENTRY2METHOD
-            WHERE ENTRY_AC = :1 
-            AND METHOD_AC = :2
-            """,
-            (e_acc, s_acc)
-        )
-    except DatabaseError:
-        cur.close()
-        return jsonify({
-            "status": False,
-            "title": "Database error",
-            "message": "Could not unintegrated "
-                       "{} from {}".format(s_acc, e_acc)
-        }), 500
-
-    cur.execute(
-        """
-        SELECT COUNT(*)
-        FROM INTERPRO.ENTRY2METHOD
-        WHERE ENTRY_AC  =:1
-        """,
-        (e_acc,)
-    )
-
-    unchecked = False
-    if not cur.fetchone()[0]:
-        # Need to uncheck the entry
-        try:
-            cur.execute(
-                """
-                UPDATE INTERPRO.ENTRY
-                SET CHECKED = 'N'
-                WHERE ENTRY_AC = :1
-                """,
-                (e_acc,)
-            )
-        except DatabaseError:
-            con.rollback()  # rollback pending transactions
-            cur.close()
-            return jsonify({
-                "status": False,
-                "title": "Database error",
-                "message": "Could not uncheck {}".format(e_acc)
-            }), 500
-        else:
-            unchecked = True
-
-    con.commit()
-    cur.close()
-    return jsonify({
-        "status": True,
-        "unchecked": unchecked
-    }), 200
-
-
-@app.route("/api/entry/<accession>/signatures/")
-def get_entry_signatures(accession):
-    cur = db.get_oracle().cursor()
-    cur.execute(
-        """
-        SELECT
-          DBCODE,
-          METHOD_AC,
-          NAME,
-          PROTEIN_COUNT
-        FROM {}.METHOD
-        WHERE METHOD_AC IN (
-          SELECT METHOD_AC
-          FROM INTERPRO.ENTRY2METHOD
-          WHERE ENTRY_AC = :1
-        )
-        ORDER BY METHOD_AC
-        """.format(app.config['DB_SCHEMA']),
-        (accession,)
-    )
-
-    signatures = []
-    for row in cur:
-        database = xref.find_ref(row[0], row[1])
-
-        signatures.append({
-            "accession": row[1],
-            "name": row[2],
-            "num_proteins": row[3],
-            "link": database.gen_link(),
-            "color": database.color,
-            "database": database.name
-        })
-
-    cur.close()
-    return jsonify(signatures), 200
-
-
-@app.route("/api/entry/<accession>/signatures/annotations/")
-def get_entry_signatures_annotations(accession):
-    cur = db.get_oracle().cursor()
-    cur.execute(
-        """
-        SELECT
-          METHOD_AC,
-          NAME,
-          ABSTRACT,
-          ABSTRACT_LONG
-        FROM {}.METHOD
-        WHERE METHOD_AC IN (
-          SELECT METHOD_AC
-          FROM INTERPRO.ENTRY2METHOD
-          WHERE ENTRY_AC = :1
-        )
-        ORDER BY METHOD_AC
-        """.format(app.config['DB_SCHEMA']),
-        (accession,)
-    )
-
-    signatures = []
-    for row in cur:
-        if row[2] is not None:
-            text = row[2]
-        elif row[3] is not None:
-            text = row[3].read()  # CLOB object
-        else:
-            text = None
-
-        signatures.append({
-            "accession": row[0],
-            "name": row[1],
-            "text": text
-        })
-
-    cur.close()
-    return jsonify(signatures), 200
